added varialbes for gyro exponent and scale on settingsDisplayer. Still need to implement toggling of views (views of the GUI to show settings, profiles, etc). Also profiled code and determined that the main while loop in MidasMain is the CPU hog and requires a sleep statement.

+++ PMMain/Midas/MidasGUI/MainGUI.cpp
@@ -210,8 +210,8 @@ void MainGUI::connectSignallerToSettingsDisplayer(SettingsSignaller *signaller)
         QObject::connect(settingsDisplayer, SIGNAL(emitBuzzFeedbackChange(unsigned int)),
             signaller, SLOT(handleBuzzFeedbackChange(unsigned int)));
 
-        QObject::connect(settingsDisplayer, SIGNAL(emitGyroPowerValue(double)),
-            signaller, SLOT(hanldeGyroPowerValueChanged(double)));
+        QObject::connect(settingsDisplayer, SIGNAL(emitGyroPowerValue(int)),
+            signaller, SLOT(hanldeGyroPowerValueChanged(int)));
         QObject::connect(settingsDisplayer, SIGNAL(emitGyroScaleDownValue(double)),
             signaller, SLOT(hanldeGyroScaleDownValueChanged(double)));
     }
